Add feature to retrieve user and display his information
