+++ app/src/androidTest/java/ch/epfl/sdp/blindly/main_screen/MatchPageTest.kt
@@ -1,0 +1,4 @@
+package ch.epfl.sdp.blindly.main_screen
+
+class MatchPageTest {
+}
