+++ app/src/main/java/ch/epfl/sdp/blindly/main_screen/ViewPagerAdapter.kt
@@ -4,10 +4,16 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-
+/**
+ * This ViewPagerAdapter is used by the ViewPager in the MainScreen
+ * to instantiate the fragments
+ */
 class ViewPagerAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
 
+    /**
+     * Given a position, create the corresponding fragment
+     */
     override fun createFragment(position: Int): Fragment {
         return when(position) {
             0 -> {
