+++ app/src/main/java/ch/epfl/sdp/blindly/profile_setup/ProfilePassions.kt
@@ -8,13 +8,12 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import ch.epfl.sdp.blindly.R
 import ch.epfl.sdp.blindly.location.AndroidLocationService
+import ch.epfl.sdp.blindly.main_screen.MainScreen
 import ch.epfl.sdp.blindly.user.UserHelper
 import com.google.android.material.chip.Chip
 import com.google.android.material.chip.ChipGroup
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
-
-const val EXTRA_PASSIONS = "passions"
 
 @AndroidEntryPoint
 class ProfilePassions : AppCompatActivity() {
@@ -62,7 +61,8 @@ class ProfilePassions : AppCompatActivity() {
             //correct numbers of selection
             else -> {
                 getCheckedChip()
-                val intent = Intent(this, ProfileAudioRecording::class.java)
+                //TODO change intent back to ProfileAudioRecording
+                val intent = Intent(this, MainScreen::class.java)
                 setUser()
                 startActivity(intent)
             }
