+++ app/src/main/java/ch/epfl/sdp/blindly/user/UserRepository.kt
@@ -14,7 +14,6 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
-
 @Module
 @InstallIn(SingletonComponent::class)
 class UserRepositoryModule {
@@ -26,25 +25,41 @@ class UserRepositoryModule {
     fun provideUserRepository(): UserRepository = UserRepository(db, userCache)
 }
 
-
+/**
+ * This is the main access point to firestore
+ */
 class UserRepository @Inject constructor(
     private val db: FirebaseFirestore,
     private val userCache: UserCache) {
 
     companion object {
-        private const val TAG = "FirebaseProfileService"
+        private const val TAG = "UserRepository"
         private const val USER_COLLECTION: String = "usersMeta"
     }
 
+    /**
+     * Given a uid, if the user is cached locally return this user, otherwise
+     * look for the user in firestore and update the cache
+     * @param uid: The uid of the user to retrieve
+     *
+     * @return the user with the corresponding uid or null if he/she/it doesn't exist
+     */
     @RequiresApi(Build.VERSION_CODES.N)
     suspend fun getUser(uid: String): User? {
         val cached: User? = userCache.get(uid)
         if (cached != null) {
+            Log.e(TAG, "Found user with uid: $uid in cache")
             return cached
         }
         return refreshUser(uid)
     }
 
+    /**
+     * Look for the user with the corresponding uid in firestore and store it in the local cache
+     * @param uid: The uid of the user to retrieve in firestore
+     *
+     * @return the user with the corresponding uid or null if he/she/it doesn't exist
+     */
     private suspend fun refreshUser(uid: String): User? {
         return try {
             val freshUser = db.collection(USER_COLLECTION)
@@ -52,6 +67,7 @@ class UserRepository @Inject constructor(
             if (freshUser != null) {
                 userCache.put(uid, freshUser)
             }
+            Log.d(TAG, "Found user with uid: $uid in firestore")
             return freshUser
         } catch (e: Exception) {
             Log.e(TAG, "Error getting user details", e)
@@ -60,10 +76,13 @@ class UserRepository @Inject constructor(
     }
 
     /**
-     * field: the field of the value to change inside the database
-     * newValue: the new value to set for the user
+     * Update a given field of the user's information (and call refreshUser to update or set the
+     * user in the local cache)
+     * @param uid: the uid of the user to update
+     * @param field: the field of the value to change inside the database
+     * @param newValue: the new value to set for the user
      */
-    /*@RequiresApi(Build.VERSION_CODES.N)
+    @RequiresApi(Build.VERSION_CODES.N)
     suspend fun <T> updateProfile(uid: String, field: String, newValue: T) {
         if(newValue !is String || newValue !is ArrayList<*>)
             throw IllegalArgumentException("Expected String or ArrayList<String>")
@@ -71,5 +90,5 @@ class UserRepository @Inject constructor(
             .document(uid)
             .update(field, newValue)
         refreshUser(uid)
-    }*/
+    }
 }
