+++ app/src/test/java/ch/epfl/sdp/blindly/UserCacheTest.kt
@@ -1,0 +1,5 @@
+package ch.epfl.sdp.blindly
+
+class UserCacheTest {
+
+}
